Separate stories from the index.ts
